pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "outbound-feed-monitor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.100.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
outbound-feed-monitor = "monitor:main"

[tool.setuptools]
py-modules = ["monitor"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]
